10_standalone_fp8_replicate_based/requirements.txt
# Core PyTorch and CUDA dependencies
torch==2.7.0
torchvision
torchaudio
triton

# NVIDIA CUDA libraries
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12

# ML and AI libraries
transformers==4.43.3
diffusers==0.32.2
huggingface-hub==0.25.0
tokenizers==0.19.1
safetensors==0.4.3
accelerate==1.10.1

# Quantization and optimization
torchao==0.12.0
quanto==0.2.0

# Image processing
pillow==11.0.0
opencv-python-headless==4.10.0.84

# Numerical computing
numpy==1.26.4
einops==0.8.0

# Utilities
tqdm==4.67.1
requests==2.25.1
pydantic==2.11.7
pydantic_core==2.33.2
regex==2025.7.34
sentencepiece==0.2.0

# Web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-multipart==0.0.20
